refactoring statics
