+++ novawallet/Modules/Staking/StakingMain/NominationPools/ViewModel/StakingNominationPoolsStatics.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct StakingNominationPoolsStatics: StakingMainStaticViewModelProtocol {
+    func networkInfoTitle(for locale: Locale) -> String {
+        R.string.localizable.stakingPoolNetworkInfo(preferredLanguages: locale.rLanguages)
+    }
+}
