+++ novawallet/Modules/Staking/StakingMain/Relaychain/ViewModel/StakingRelaychainStatics.swift
@@ -1,26 +1,3 @@
 import Foundation
 
-struct StakingRelaychainStatics: StakingMainStaticViewModelProtocol {
-    func networkInfoActiveNominators(for locale: Locale) -> String {
-        R.string.localizable.stakingMainActiveNominatorsTitle(
-            preferredLanguages: locale.rLanguages
-        )
-    }
-
-    func actionsYourValidators(for locale: Locale) -> String {
-        R.string.localizable.stakingYourValidatorsTitle(
-            preferredLanguages: locale.rLanguages
-        )
-    }
-
-    func waitingNextEra(for timeString: String, locale: Locale) -> String {
-        R.string.localizable.stakingWaitingNextEraFormat(
-            timeString,
-            preferredLanguages: locale.rLanguages
-        ).uppercased()
-    }
-
-    func networkInfoTitle(for locale: Locale) -> String {
-        R.string.localizable.stakingNetworkInfoTitle(preferredLanguages: locale.rLanguages)
-    }
-}
+struct StakingRelaychainStatics: StakingMainStaticViewModelProtocol {}
